api/http: harden session limit handling
handle the session limit just once; don't
spam both ends with error messages if we already did that.

ignore evt above the session limit - do not send to user.

free the db user session as part of handling the limit.

to avoid bigger refactoring, we still rely on the
the device 'stop' handler to clean up the session map.

in general, all of this needs a bigger refactor to catch
all the corner cases.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>