Fix race condition if reader returns while writer is busy
There is a race condition if the reader returns while the websocket
writer is busy (e.g. publishing on a dead NATS subject). While this is
happening the client may reconnect and the conenction status reaches an
inconsistent state. This commit ensures that the context is properly
canceled while either handler returns, and the nats.Client.Publish takes
a context which can interrupt when the publisher is blocking for
replies.

changelog: none

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>