fix: Connection status race on rapid reconnect
Devices that reconnect too fast may result in a state where the
persisted status shows “disconnected” while the device is actually
connected. This especially happens when the network latency is very low
(< 5ms) which causes the device to establish a new connection before the
old request handler had time to mark it as closed.

Changelog: Commit
Ticket: MEN-7333
Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>